Move field display definitions into fields directory
