+++ datadoc/frontend/components/DatasetTab.py
@@ -5,12 +5,12 @@ from datadoc.frontend.Builders import (
     make_dataset_metadata_accordion_item,
     make_ssb_styled_tab,
 )
-from datadoc.frontend.DisplayDataset import (
+from datadoc.frontend.fields.DisplayDataset import (
     NON_EDITABLE_DATASET_METADATA,
     OBLIGATORY_EDITABLE_DATASET_METADATA,
     OPTIONAL_DATASET_METADATA,
 )
-from datadoc.frontend.DisplayVariables import DISPLAY_VARIABLES
+from datadoc.frontend.fields.DisplayVariables import DISPLAY_VARIABLES
 from datadoc.utils import get_display_values
 import datadoc.state as state
